requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
tortoise-orm==0.23.0
aerich==0.8.1
asyncpg==0.30.0
pydantic==2.10.4
marshmallow==3.23.2
python-dotenv==1.0.1
tomlkit==0.13.2
pytest
pytest-asyncio
httpx
redis